requirements.txt
pyyaml
requests
python-ldap